# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test ihold",
        "version": "1.0"
    },
    "paths": {
...
        "/api/merchants": {
            "get": {
                "tags": [
                    "Merchants"
                ],
                "summary": "List all merchants",
                "operationId": "29dc2feea9d224f1903cfe3d27bd81fb",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Merchants"
                ],
                "summary": "Create a merchants",
                "description": "Store a new Order.",
                "operationId": "15db992a94c43f36e50daf7853b53ca2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "merchant_name": {
                                        "type": "string",
                                        "example": "Zé delivery"
                                    },
                                    "admin_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "Merchants registered successfully"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/api/merchants/{id}": {
            "get": {
                "tags": [
                    "Merchants"
                ],
                "summary": "Get information about an merchants",
                "description": "Show a Merchants.",
                "operationId": "56beef28bbe1d805bab17fb72672742e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the merchants to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "Merchants retrieved successfully"
                            }
                        }
                    },
                    "404": {
                        "description": "Merchants not found"
                    },
                    "403": {
                        "description": "You do not have permission to retrieve the merchants"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/api/merchants/{$id}/": {
            "put": {
                "tags": [
                    "Merchants"
                ],
                "summary": "Update a Merchants",
                "description": "Update a Order.",
                "operationId": "b5d39b1615d465cf73a31877188716b3",
                "parameters": [
                    {
                        "name": "merchantId",
                        "in": "path",
                        "description": "ID of the merchants to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "merchant_name": {
                                        "type": "string",
                                        "example": "João"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "Merchants updated successfully"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Merchants"
                ],
                "summary": "Delete a merchants",
                "description": "Delete a Order.",
                "operationId": "a21e569f05d8e3276f9de39180ecfda4",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "ID of the merchants to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "Merchant deleted successfully"
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "403": {
                        "description": "You do not have permission to delete the Merchants"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}